+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Allow search by words (establishment name)
+- Allow search by GIAS establishment number
 
 ## [Release-44][release-44]